⤵️ Backport of #54006 → `stable/8.9`

relates to #54002

---
<sub><img src="https://avatars.githubusercontent.com/u/97796249?s=16"
width="16" height="16" align="absmiddle"> Created by <a
href="https://github.com/korthout/backport-action">backport-action</a></sub>

+++ optimize/backend/src/main/java/io/camunda/optimize/service/db/es/filter/util/DateHistogramFilterUtilES.java
@@ -188,19 +188,17 @@ public final class DateHistogramFilterUtilES {
                             FieldDateMath.of(
                                 f ->
                                     f.value(
-                                        (double)
-                                            start
-                                                .atZoneSameInstant(context.getTimezone())
-                                                .toInstant()
-                                                .toEpochMilli())))
+                                        start
+                                            .atZoneSameInstant(context.getTimezone())
+                                            .toInstant()
+                                            .toEpochMilli())))
                         .max(
                             FieldDateMath.of(
                                 f ->
                                     f.value(
-                                        (double)
-                                            filterEnd
-                                                .atZoneSameInstant(context.getTimezone())
-                                                .toInstant()
-                                                .toEpochMilli())))));
+                                        filterEnd
+                                            .atZoneSameInstant(context.getTimezone())
+                                            .toInstant()
+                                            .toEpochMilli())))));
   }
 }
